03Builder/Students/2019/BartoszWisniewski/builder/src/main/java/classic/Game.java
package classic;

public class Game {

	private GameType gameType;
	private int numberOfPlayers;
	private String dealer;

	public GameType getGameType() {
		return gameType;
	}

	public void setGameType(GameType gameType) {
		this.gameType = gameType;
	}

	public int getNumberOfPlayers() {
		return numberOfPlayers;
	}

	public void setNumberOfPlayers(int numberOfPlayers) {
		this.numberOfPlayers = numberOfPlayers;
	}

	public String getDealer() {
		return dealer;
	}

	public void setDealer(String dealer) {
		this.dealer = dealer;
	}

}
